*Added all the files I altered and created on my Linux.
*Setup.py has the new Property Self.load_model, if set to True main.py will
look for an already created network in the checkpoint_dir

*getimages.py has been added. It requires Selenium to be installed and i
have only tested it with the Chromedriver. You define the Search Term
and number of Images you want to check at the bottom (I will improve
this later) and then get the choice to save each image in the "yes" or
"no" Folder.

*Added "import glob" to gifmaker.py, it didn't work on linux otherwise

*loadnetwork.py has been added. It loads the Generator network from the
checkpoint directory and uses it to generate images.

*main.py now saves the generator and discriminator network every time it
saves an image. I have set this to occur every 100 Epochs. (this
interval is abitrary and can be changed at any time.)